backend/requirements_complete.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Security & Crypto
cryptography==41.0.7  # للـ HMAC verification

# Date/Time
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10  # أسرع من json العادي

# Logging (optional but recommended)
structlog==23.2.0  # Better logging

# CLI (for manage.py scripts)
click==8.1.7  # For CLI commands
rich==13.7.0  # Pretty terminal output

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0  # للـ mocking
faker==20.1.0  # للـ test data generation

# Code Quality (optional - development)
black==23.12.1  # Code formatter
ruff==0.1.8  # Fast linter
mypy==1.7.1  # Type checking

# Process Management (optional - للـ production)
# pm2 (install via npm: npm install -g pm2)

# Monitoring (optional - للـ production)
# prometheus-client==0.19.0  # Metrics
# sentry-sdk[fastapi]==1.39.1  # Error tracking
